requirements.txt
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dateutils==0.6.12
decorator==4.0.11
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
joblib==1.3.2
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
praw @ git+https://github.com/praw-dev/praw.git@f9bc8044e07cf651e12de8509fc6c7d57da1b7dc
prawcore==2.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.10.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
tweety-ns==1.0.9.8
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.7.0
